+++ astlib/migrate_504_503.ml
@@ -2,9 +2,17 @@ open Stdlib0
 module From = Ast_504
 module To = Ast_503
 
-let migration_error loc missing_feature =
+module Bivariant_param = struct
+  let list_map ~f l =
+    List.fold_left_map ~acc:false l ~f:(fun ~acc x ->
+        let contains_bivariant, x' = f x in
+        (acc || contains_bivariant, x'))
+end
+
+let bivariant_error ~loc =
   Location.raise_errorf ~loc
-    "migration error: %s are not supported before OCaml 5.4" missing_feature
+    "Ppxlib migration error: bivariant type parameters cannot be migrated from \
+     OCaml 5.4 to 5.3"
 
 let rec copy_toplevel_phrase :
     Ast_504.Parsetree.toplevel_phrase -> Ast_503.Parsetree.toplevel_phrase =
@@ -528,12 +536,14 @@ and copy_structure : Ast_504.Parsetree.structure -> Ast_503.Parsetree.structure
 and copy_structure_item :
     Ast_504.Parsetree.structure_item -> Ast_503.Parsetree.structure_item =
  fun { Ast_504.Parsetree.pstr_desc; Ast_504.Parsetree.pstr_loc } ->
+  let loc = copy_location pstr_loc in
   {
-    Ast_503.Parsetree.pstr_desc = copy_structure_item_desc pstr_desc;
-    Ast_503.Parsetree.pstr_loc = copy_location pstr_loc;
+    Ast_503.Parsetree.pstr_desc =
+      copy_structure_item_desc_with_loc ~loc pstr_desc;
+    Ast_503.Parsetree.pstr_loc = loc;
   }
 
-and copy_structure_item_desc :
+and copy_structure_item_desc_with_loc ~loc :
     Ast_504.Parsetree.structure_item_desc ->
     Ast_503.Parsetree.structure_item_desc = function
   | Ast_504.Parsetree.Pstr_eval (x0, x1) ->
@@ -544,10 +554,18 @@ and copy_structure_item_desc :
   | Ast_504.Parsetree.Pstr_primitive x0 ->
       Ast_503.Parsetree.Pstr_primitive (copy_value_description x0)
   | Ast_504.Parsetree.Pstr_type (x0, x1) ->
-      Ast_503.Parsetree.Pstr_type
-        (copy_rec_flag x0, List.map copy_type_declaration x1)
+      let rec_flag = copy_rec_flag x0 in
+      let contains_bivariant, tds =
+        Bivariant_param.list_map ~f:copy_type_declaration_ x1
+      in
+      if contains_bivariant then
+        Encoding_504.To_503.encode_bivariant_pstr_type ~loc rec_flag tds
+      else Ast_503.Parsetree.Pstr_type (rec_flag, tds)
   | Ast_504.Parsetree.Pstr_typext x0 ->
-      Ast_503.Parsetree.Pstr_typext (copy_type_extension x0)
+      let contains_bivariant, ty_ext = copy_type_extension_ x0 in
+      if contains_bivariant then
+        Encoding_504.To_503.encode_bivariant_pstr_typext ~loc ty_ext
+      else Ast_503.Parsetree.Pstr_typext ty_ext
   | Ast_504.Parsetree.Pstr_exception x0 ->
       Ast_503.Parsetree.Pstr_exception (copy_type_exception x0)
   | Ast_504.Parsetree.Pstr_module x0 ->
@@ -559,10 +577,19 @@ and copy_structure_item_desc :
   | Ast_504.Parsetree.Pstr_open x0 ->
       Ast_503.Parsetree.Pstr_open (copy_open_declaration x0)
   | Ast_504.Parsetree.Pstr_class x0 ->
-      Ast_503.Parsetree.Pstr_class (List.map copy_class_declaration x0)
+      let contains_bivariant, cds =
+        Bivariant_param.list_map ~f:copy_class_declaration_ x0
+      in
+      if contains_bivariant then
+        Encoding_504.To_503.encode_bivariant_pstr_class ~loc cds
+      else Ast_503.Parsetree.Pstr_class cds
   | Ast_504.Parsetree.Pstr_class_type x0 ->
-      Ast_503.Parsetree.Pstr_class_type
-        (List.map copy_class_type_declaration x0)
+      let contains_bivariant, ctds =
+        Bivariant_param.list_map ~f:copy_class_type_declaration_ x0
+      in
+      if contains_bivariant then
+        Encoding_504.To_503.encode_bivariant_pstr_class_type ~loc ctds
+      else Ast_503.Parsetree.Pstr_class_type ctds
   | Ast_504.Parsetree.Pstr_include x0 ->
       Ast_503.Parsetree.Pstr_include (copy_include_declaration x0)
   | Ast_504.Parsetree.Pstr_attribute x0 ->
@@ -575,9 +602,10 @@ and copy_include_declaration :
     Ast_503.Parsetree.include_declaration =
  fun x -> copy_include_infos copy_module_expr x
 
-and copy_class_declaration :
-    Ast_504.Parsetree.class_declaration -> Ast_503.Parsetree.class_declaration =
- fun x -> copy_class_infos copy_class_expr x
+and copy_class_declaration_ :
+    Ast_504.Parsetree.class_declaration ->
+    bool * Ast_503.Parsetree.class_declaration =
+ fun x -> copy_class_infos_ copy_class_expr x
 
 and copy_class_expr :
     Ast_504.Parsetree.class_expr -> Ast_503.Parsetree.class_expr =
@@ -750,13 +778,14 @@ and copy_module_type :
        Ast_504.Parsetree.pmty_loc;
        Ast_504.Parsetree.pmty_attributes;
      } ->
+  let loc = copy_location pmty_loc in
   {
-    Ast_503.Parsetree.pmty_desc = copy_module_type_desc pmty_desc;
-    Ast_503.Parsetree.pmty_loc = copy_location pmty_loc;
+    Ast_503.Parsetree.pmty_desc = copy_module_type_desc_with_loc ~loc pmty_desc;
+    Ast_503.Parsetree.pmty_loc = loc;
     Ast_503.Parsetree.pmty_attributes = copy_attributes pmty_attributes;
   }
 
-and copy_module_type_desc :
+and copy_module_type_desc_with_loc ~loc :
     Ast_504.Parsetree.module_type_desc -> Ast_503.Parsetree.module_type_desc =
   function
   | Ast_504.Parsetree.Pmty_ident x0 ->
@@ -767,8 +796,13 @@ and copy_module_type_desc :
       Ast_503.Parsetree.Pmty_functor
         (copy_functor_parameter x0, copy_module_type x1)
   | Ast_504.Parsetree.Pmty_with (x0, x1) ->
-      Ast_503.Parsetree.Pmty_with
-        (copy_module_type x0, List.map copy_with_constraint x1)
+      let mty = copy_module_type x0 in
+      let contains_bivariant, constraints =
+        Bivariant_param.list_map ~f:copy_with_constraint_ x1
+      in
+      if contains_bivariant then
+        Encoding_504.To_503.encode_bivariant_pmty_with ~loc mty constraints
+      else Ast_503.Parsetree.Pmty_with (copy_module_type x0, constraints)
   | Ast_504.Parsetree.Pmty_typeof x0 ->
       Ast_503.Parsetree.Pmty_typeof (copy_module_expr x0)
   | Ast_504.Parsetree.Pmty_extension x0 ->
@@ -776,27 +810,33 @@ and copy_module_type_desc :
   | Ast_504.Parsetree.Pmty_alias x0 ->
       Ast_503.Parsetree.Pmty_alias (copy_loc copy_Longident_t x0)
 
-and copy_with_constraint :
-    Ast_504.Parsetree.with_constraint -> Ast_503.Parsetree.with_constraint =
-  function
+and copy_with_constraint_ :
+    Ast_504.Parsetree.with_constraint ->
+    bool * Ast_503.Parsetree.with_constraint = function
   | Ast_504.Parsetree.Pwith_type (x0, x1) ->
-      Ast_503.Parsetree.Pwith_type
-        (copy_loc copy_Longident_t x0, copy_type_declaration x1)
+      let lident_loc = copy_loc copy_Longident_t x0 in
+      let contains_bivariant, td = copy_type_declaration_ x1 in
+      (contains_bivariant, Ast_503.Parsetree.Pwith_type (lident_loc, td))
   | Ast_504.Parsetree.Pwith_module (x0, x1) ->
-      Ast_503.Parsetree.Pwith_module
-        (copy_loc copy_Longident_t x0, copy_loc copy_Longident_t x1)
+      ( false,
+        Ast_503.Parsetree.Pwith_module
+          (copy_loc copy_Longident_t x0, copy_loc copy_Longident_t x1) )
   | Ast_504.Parsetree.Pwith_modtype (x0, x1) ->
-      Ast_503.Parsetree.Pwith_modtype
-        (copy_loc copy_Longident_t x0, copy_module_type x1)
+      ( false,
+        Ast_503.Parsetree.Pwith_modtype
+          (copy_loc copy_Longident_t x0, copy_module_type x1) )
   | Ast_504.Parsetree.Pwith_modtypesubst (x0, x1) ->
-      Ast_503.Parsetree.Pwith_modtypesubst
-        (copy_loc copy_Longident_t x0, copy_module_type x1)
+      ( false,
+        Ast_503.Parsetree.Pwith_modtypesubst
+          (copy_loc copy_Longident_t x0, copy_module_type x1) )
   | Ast_504.Parsetree.Pwith_typesubst (x0, x1) ->
-      Ast_503.Parsetree.Pwith_typesubst
-        (copy_loc copy_Longident_t x0, copy_type_declaration x1)
+      let lident_loc = copy_loc copy_Longident_t x0 in
+      let contains_bivariant, td = copy_type_declaration_ x1 in
+      (contains_bivariant, Ast_503.Parsetree.Pwith_typesubst (lident_loc, td))
   | Ast_504.Parsetree.Pwith_modsubst (x0, x1) ->
-      Ast_503.Parsetree.Pwith_modsubst
-        (copy_loc copy_Longident_t x0, copy_loc copy_Longident_t x1)
+      ( false,
+        Ast_503.Parsetree.Pwith_modsubst
+          (copy_loc copy_Longident_t x0, copy_loc copy_Longident_t x1) )
 
 and copy_signature : Ast_504.Parsetree.signature -> Ast_503.Parsetree.signature
     =
@@ -805,23 +845,38 @@ and copy_signature : Ast_504.Parsetree.signature -> Ast_503.Parsetree.signature
 and copy_signature_item :
     Ast_504.Parsetree.signature_item -> Ast_503.Parsetree.signature_item =
  fun { Ast_504.Parsetree.psig_desc; Ast_504.Parsetree.psig_loc } ->
+  let loc = copy_location psig_loc in
   {
-    Ast_503.Parsetree.psig_desc = copy_signature_item_desc psig_desc;
-    Ast_503.Parsetree.psig_loc = copy_location psig_loc;
+    Ast_503.Parsetree.psig_desc =
+      copy_signature_item_desc_with_loc ~loc psig_desc;
+    Ast_503.Parsetree.psig_loc = loc;
   }
 
-and copy_signature_item_desc :
+and copy_signature_item_desc_with_loc ~loc :
     Ast_504.Parsetree.signature_item_desc ->
     Ast_503.Parsetree.signature_item_desc = function
   | Ast_504.Parsetree.Psig_value x0 ->
       Ast_503.Parsetree.Psig_value (copy_value_description x0)
   | Ast_504.Parsetree.Psig_type (x0, x1) ->
-      Ast_503.Parsetree.Psig_type
-        (copy_rec_flag x0, List.map copy_type_declaration x1)
+      let rec_flag = copy_rec_flag x0 in
+      let contains_bivariant, tds =
+        Bivariant_param.list_map ~f:copy_type_declaration_ x1
+      in
+      if contains_bivariant then
+        Encoding_504.To_503.encode_bivariant_psig_type ~loc rec_flag tds
+      else Ast_503.Parsetree.Psig_type (rec_flag, tds)
   | Ast_504.Parsetree.Psig_typesubst x0 ->
-      Ast_503.Parsetree.Psig_typesubst (List.map copy_type_declaration x0)
+      let contains_bivariant, tds =
+        Bivariant_param.list_map ~f:copy_type_declaration_ x0
+      in
+      if contains_bivariant then
+        Encoding_504.To_503.encode_bivariant_psig_typesubst ~loc tds
+      else Ast_503.Parsetree.Psig_typesubst tds
   | Ast_504.Parsetree.Psig_typext x0 ->
-      Ast_503.Parsetree.Psig_typext (copy_type_extension x0)
+      let contains_bivariant, ty_ext = copy_type_extension_ x0 in
+      if contains_bivariant then
+        Encoding_504.To_503.encode_bivariant_psig_typext ~loc ty_ext
+      else Ast_503.Parsetree.Psig_typext ty_ext
   | Ast_504.Parsetree.Psig_exception x0 ->
       Ast_503.Parsetree.Psig_exception (copy_type_exception x0)
   | Ast_504.Parsetree.Psig_module x0 ->
@@ -839,23 +894,33 @@ and copy_signature_item_desc :
   | Ast_504.Parsetree.Psig_include x0 ->
       Ast_503.Parsetree.Psig_include (copy_include_description x0)
   | Ast_504.Parsetree.Psig_class x0 ->
-      Ast_503.Parsetree.Psig_class (List.map copy_class_description x0)
+      let contains_bivariant, cds =
+        Bivariant_param.list_map ~f:copy_class_description_ x0
+      in
+      if contains_bivariant then
+        Encoding_504.To_503.encode_bivariant_psig_class ~loc cds
+      else Ast_503.Parsetree.Psig_class cds
   | Ast_504.Parsetree.Psig_class_type x0 ->
-      Ast_503.Parsetree.Psig_class_type
-        (List.map copy_class_type_declaration x0)
+      let contains_bivariant, ctds =
+        Bivariant_param.list_map ~f:copy_class_type_declaration_ x0
+      in
+      if contains_bivariant then
+        Encoding_504.To_503.encode_bivariant_psig_class_type ~loc ctds
+      else Ast_503.Parsetree.Psig_class_type ctds
   | Ast_504.Parsetree.Psig_attribute x0 ->
       Ast_503.Parsetree.Psig_attribute (copy_attribute x0)
   | Ast_504.Parsetree.Psig_extension (x0, x1) ->
       Ast_503.Parsetree.Psig_extension (copy_extension x0, copy_attributes x1)
 
-and copy_class_type_declaration :
+and copy_class_type_declaration_ :
     Ast_504.Parsetree.class_type_declaration ->
-    Ast_503.Parsetree.class_type_declaration =
- fun x -> copy_class_infos copy_class_type x
+    bool * Ast_503.Parsetree.class_type_declaration =
+ fun x -> copy_class_infos_ copy_class_type x
 
-and copy_class_description :
-    Ast_504.Parsetree.class_description -> Ast_503.Parsetree.class_description =
- fun x -> copy_class_infos copy_class_type x
+and copy_class_description_ :
+    Ast_504.Parsetree.class_description ->
+    bool * Ast_503.Parsetree.class_description =
+ fun x -> copy_class_infos_ copy_class_type x
 
 and copy_class_type :
     Ast_504.Parsetree.class_type -> Ast_503.Parsetree.class_type =
@@ -941,11 +1006,11 @@ and copy_extension : Ast_504.Parsetree.extension -> Ast_503.Parsetree.extension
   let x0, x1 = x in
   (copy_loc (fun x -> x) x0, copy_payload x1)
 
-and copy_class_infos :
+and copy_class_infos_ :
     'f0 'g0.
     ('f0 -> 'g0) ->
     'f0 Ast_504.Parsetree.class_infos ->
-    'g0 Ast_503.Parsetree.class_infos =
+    bool * 'g0 Ast_503.Parsetree.class_infos =
  fun f0
      {
        Ast_504.Parsetree.pci_virt;
@@ -955,21 +1020,18 @@ and copy_class_infos :
        Ast_504.Parsetree.pci_loc;
        Ast_504.Parsetree.pci_attributes;
      } ->
-  {
-    Ast_503.Parsetree.pci_virt = copy_virtual_flag pci_virt;
-    Ast_503.Parsetree.pci_params =
-      List.map
-        (fun x ->
-          let x0, x1 = x in
-          ( copy_core_type x0,
-            let x0, x1 = x1 in
-            (copy_variance x0, copy_injectivity x1) ))
-        pci_params;
-    Ast_503.Parsetree.pci_name = copy_loc (fun x -> x) pci_name;
-    Ast_503.Parsetree.pci_expr = f0 pci_expr;
-    Ast_503.Parsetree.pci_loc = copy_location pci_loc;
-    Ast_503.Parsetree.pci_attributes = copy_attributes pci_attributes;
-  }
+  let contains_bivariant, params = copy_type_params pci_params in
+  let ci =
+    {
+      Ast_503.Parsetree.pci_virt = copy_virtual_flag pci_virt;
+      Ast_503.Parsetree.pci_params = params;
+      Ast_503.Parsetree.pci_name = copy_loc (fun x -> x) pci_name;
+      Ast_503.Parsetree.pci_expr = f0 pci_expr;
+      Ast_503.Parsetree.pci_loc = copy_location pci_loc;
+      Ast_503.Parsetree.pci_attributes = copy_attributes pci_attributes;
+    }
+  in
+  (contains_bivariant, ci)
 
 and copy_virtual_flag :
     Ast_504.Asttypes.virtual_flag -> Ast_503.Asttypes.virtual_flag = function
@@ -1089,8 +1151,9 @@ and copy_type_exception :
     Ast_503.Parsetree.ptyexn_attributes = copy_attributes ptyexn_attributes;
   }
 
-and copy_type_extension :
-    Ast_504.Parsetree.type_extension -> Ast_503.Parsetree.type_extension =
+and copy_type_extension_ :
+    Ast_504.Parsetree.type_extension -> bool * Ast_503.Parsetree.type_extension
+    =
  fun {
        Ast_504.Parsetree.ptyext_path;
        Ast_504.Parsetree.ptyext_params;
@@ -1099,22 +1162,19 @@ and copy_type_extension :
        Ast_504.Parsetree.ptyext_loc;
        Ast_504.Parsetree.ptyext_attributes;
      } ->
-  {
-    Ast_503.Parsetree.ptyext_path = copy_loc copy_Longident_t ptyext_path;
-    Ast_503.Parsetree.ptyext_params =
-      List.map
-        (fun x ->
-          let x0, x1 = x in
-          ( copy_core_type x0,
-            let x0, x1 = x1 in
-            (copy_variance x0, copy_injectivity x1) ))
-        ptyext_params;
-    Ast_503.Parsetree.ptyext_constructors =
-      List.map copy_extension_constructor ptyext_constructors;
-    Ast_503.Parsetree.ptyext_private = copy_private_flag ptyext_private;
-    Ast_503.Parsetree.ptyext_loc = copy_location ptyext_loc;
-    Ast_503.Parsetree.ptyext_attributes = copy_attributes ptyext_attributes;
-  }
+  let contains_bivariant, params = copy_type_params ptyext_params in
+  let te =
+    {
+      Ast_503.Parsetree.ptyext_path = copy_loc copy_Longident_t ptyext_path;
+      Ast_503.Parsetree.ptyext_params = params;
+      Ast_503.Parsetree.ptyext_constructors =
+        List.map copy_extension_constructor ptyext_constructors;
+      Ast_503.Parsetree.ptyext_private = copy_private_flag ptyext_private;
+      Ast_503.Parsetree.ptyext_loc = copy_location ptyext_loc;
+      Ast_503.Parsetree.ptyext_attributes = copy_attributes ptyext_attributes;
+    }
+  in
+  (contains_bivariant, te)
 
 and copy_extension_constructor :
     Ast_504.Parsetree.extension_constructor ->
@@ -1143,8 +1203,18 @@ and copy_extension_constructor_kind :
   | Ast_504.Parsetree.Pext_rebind x0 ->
       Ast_503.Parsetree.Pext_rebind (copy_loc copy_Longident_t x0)
 
-and copy_type_declaration :
-    Ast_504.Parsetree.type_declaration -> Ast_503.Parsetree.type_declaration =
+and copy_type_params params =
+  Bivariant_param.list_map params ~f:(fun (typ, (variance, injectivity)) ->
+      let typ' = copy_core_type typ in
+      let injectivity' = copy_injectivity injectivity in
+      match variance with
+      | Ast_504.Asttypes.Bivariant ->
+          (true, Encoding_504.To_503.encode_bivariant_param typ' injectivity')
+      | _ -> (false, (typ', (copy_variance variance, injectivity'))))
+
+and copy_type_declaration_ :
+    Ast_504.Parsetree.type_declaration ->
+    bool * Ast_503.Parsetree.type_declaration =
  fun {
        Ast_504.Parsetree.ptype_name;
        Ast_504.Parsetree.ptype_params;
@@ -1155,28 +1225,26 @@ and copy_type_declaration :
        Ast_504.Parsetree.ptype_attributes;
        Ast_504.Parsetree.ptype_loc;
      } ->
-  {
-    Ast_503.Parsetree.ptype_name = copy_loc (fun x -> x) ptype_name;
-    Ast_503.Parsetree.ptype_params =
-      List.map
-        (fun x ->
-          let x0, x1 = x in
-          ( copy_core_type x0,
-            let x0, x1 = x1 in
-            (copy_variance x0, copy_injectivity x1) ))
-        ptype_params;
-    Ast_503.Parsetree.ptype_cstrs =
-      List.map
-        (fun x ->
-          let x0, x1, x2 = x in
-          (copy_core_type x0, copy_core_type x1, copy_location x2))
-        ptype_cstrs;
-    Ast_503.Parsetree.ptype_kind = copy_type_kind ptype_kind;
-    Ast_503.Parsetree.ptype_private = copy_private_flag ptype_private;
-    Ast_503.Parsetree.ptype_manifest = Option.map copy_core_type ptype_manifest;
-    Ast_503.Parsetree.ptype_attributes = copy_attributes ptype_attributes;
-    Ast_503.Parsetree.ptype_loc = copy_location ptype_loc;
-  }
+  let contains_bivariant, params = copy_type_params ptype_params in
+  let td =
+    {
+      Ast_503.Parsetree.ptype_name = copy_loc (fun x -> x) ptype_name;
+      Ast_503.Parsetree.ptype_params = params;
+      Ast_503.Parsetree.ptype_cstrs =
+        List.map
+          (fun x ->
+            let x0, x1, x2 = x in
+            (copy_core_type x0, copy_core_type x1, copy_location x2))
+          ptype_cstrs;
+      Ast_503.Parsetree.ptype_kind = copy_type_kind ptype_kind;
+      Ast_503.Parsetree.ptype_private = copy_private_flag ptype_private;
+      Ast_503.Parsetree.ptype_manifest =
+        Option.map copy_core_type ptype_manifest;
+      Ast_503.Parsetree.ptype_attributes = copy_attributes ptype_attributes;
+      Ast_503.Parsetree.ptype_loc = copy_location ptype_loc;
+    }
+  in
+  (contains_bivariant, td)
 
 and copy_private_flag :
     Ast_504.Asttypes.private_flag -> Ast_503.Asttypes.private_flag = function
@@ -1253,7 +1321,10 @@ and copy_variance : Ast_504.Asttypes.variance -> Ast_503.Asttypes.variance =
   | Ast_504.Asttypes.Covariant -> Ast_503.Asttypes.Covariant
   | Ast_504.Asttypes.Contravariant -> Ast_503.Asttypes.Contravariant
   | Ast_504.Asttypes.NoVariance -> Ast_503.Asttypes.NoVariance
-  | Ast_504.Asttypes.Bivariant -> migration_error Location.none "bivariance"
+  | Ast_504.Asttypes.Bivariant ->
+      (* note that this should never be reached, unless a user explicitly
+       calls this very function *)
+      bivariant_error ~loc:Location.none
 
 and copy_value_description :
     Ast_504.Parsetree.value_description -> Ast_503.Parsetree.value_description =
@@ -1327,3 +1398,55 @@ and copy_loc :
   { Ast_503.Asttypes.txt = f0 txt; Ast_503.Asttypes.loc = copy_location loc }
 
 and copy_location : Location.t -> Location.t = fun x -> x
+
+(** The functions below are provided to keep a coherent and stable API with
+    other migrate_X_Y modules while allowing us to locally define variations of
+    those functions used to encode 5.4 features into the 5.3 AST. *)
+
+let copy_structure_item_desc stri_d =
+  copy_structure_item_desc_with_loc ~loc:Location.none stri_d
+
+let copy_module_type_desc pmty =
+  copy_module_type_desc_with_loc ~loc:Location.none pmty
+
+let copy_signature_item_desc sigi_desc =
+  copy_signature_item_desc_with_loc ~loc:Location.none sigi_desc
+
+let copy_class_infos :
+    'f0 'g0.
+    ('f0 -> 'g0) ->
+    'f0 Ast_504.Parsetree.class_infos ->
+    'g0 Ast_503.Parsetree.class_infos =
+ fun f0 ci ->
+  let contains_bivariant, ci = copy_class_infos_ f0 ci in
+  if contains_bivariant then bivariant_error ~loc:ci.Ast_503.Parsetree.pci_loc
+  else ci
+
+let copy_class_declaration cd =
+  let contains_bivariant, cd' = copy_class_declaration_ cd in
+  if contains_bivariant then bivariant_error ~loc:cd.pci_loc else cd'
+
+let copy_with_constraint wc =
+  let contains_bivariant, wc' = copy_with_constraint_ wc in
+  if contains_bivariant then
+    match wc with
+    | Pwith_type (_, td) | Pwith_typesubst (_, td) ->
+        bivariant_error ~loc:td.ptype_loc
+    | _ -> assert false
+  else wc'
+
+let copy_class_type_declaration ctd =
+  let contains_bivariant, ctd' = copy_class_type_declaration_ ctd in
+  if contains_bivariant then bivariant_error ~loc:ctd.pci_loc else ctd'
+
+let copy_class_description cd =
+  let contains_bivariant, cd' = copy_class_description_ cd in
+  if contains_bivariant then bivariant_error ~loc:cd.pci_loc else cd'
+
+let copy_type_extension te =
+  let contains_bivariant, te' = copy_type_extension_ te in
+  if contains_bivariant then bivariant_error ~loc:te.ptyext_loc else te'
+
+let copy_type_declaration td =
+  let contains_bivariant, td' = copy_type_declaration_ td in
+  if contains_bivariant then bivariant_error ~loc:td.ptype_loc else td'